requirements.txt
absl-py
altgraph
annotated-types
APScheduler
astunparse
async-timeout
attrs
blinker
blis
build
cachelib
catalogue
certifi
cffi
charset-normalizer
click
cloudpathlib
colorama
coloredlogs
confection
contourpy
cryptography
cycler
cymem
decorator
docutils
elastic-transport
elasticsearch
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
<<<<<<< HEAD
filelock
Flask
Flask-Avatars
Flask-Caching
Flask-Login
Flask-Moment
Flask-SQLAlchemy
flatbuffers
fonttools
fsspec
future
gast
geocoder
google-pasta
greenlet
grpcio
h5py
huggingface-hub
humanfriendly
idna
imageio
importlib_metadata
iteration_utilities
itsdangerous
jaraco.classes
jaraco.context
jaraco.functools
jeepney
Jinja2
joblib
jsonschema
jsonschema-specifications
keras
keyring
kiwisolver
langcodes
language_data
lazy_loader
libclang
llvmlite
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml-dtypes
more-itertools
mpmath
murmurhash
mysql-connector-python
namex
networkx
nh3
nltk
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
onnxruntime
opencv-python
opencv-python-headless
opt_einsum
optree
packaging
pandas
pillow
pkginfo
platformdirs
pooch
preshed
protobuf
psutil
py-cpuinfo
pycparser
pydantic
pydantic_core
Pygments
pyinstaller
pyinstaller-hooks-contrib
PyMatting
PyMySQL
pyparsing
pyproject_hooks
python-dateutil
python-decouple
python-dotenv
pytz
PyYAML
ratelim
readme_renderer
redis
referencing
regex
rembg
requests
requests-toolbelt
rfc3986
rich
rpds-py
safetensors
scikit-image
scikit-learn
scipy
seaborn
SecretStorage
sentence-transformers
sentencepiece
setuptools
shellingham
six
smart-open
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-intel
tensorflow-io-gcs-filesystem
tensorflow_cpu
termcolor
thinc
threadpoolctl
tifffile
tokenizers
torch
torchvision
tqdm
transformers
triton
twine
typer
typing_extensions
tzdata
tzlocal
ultralytics
ultralytics-thop
urllib3
wasabi
weasel
Werkzeug
wheel
wrapt
zipp
=======
filelock==3.16.1
Flask==3.0.3
Flask-Avatars==0.2.3
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
geocoder==1.38.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.0
h5py==3.12.1
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
keyring==25.4.1
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
murmurhash==1.0.10
mysql-connector-python==9.1.0
namex==0.0.8
networkx==3.4.1
nh3==0.2.18
numba==0.60.0
numpy
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pkginfo==1.10.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
protobuf==4.25.5
psutil==6.1.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyMatting==1.1.12
PyMySQL==1.1.1
pyparsing==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
ratelim==0.1.6
readme_renderer==44.0
redis==5.1.1
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_cpu==2.18.0
termcolor==2.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
transformers==4.46.0
triton==3.1.0
twine==5.1.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.15
ultralytics-thop==2.0.9
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
zipp==3.20.2
>>>>>>> f2bc93d (update)
